+++ tests/config/test_autoconfig.py
@@ -264,17 +264,18 @@ class TestAutoConfigResolver:
     @pytest.mark.parametrize(
         "max_seq_input, expected_max_seq",
         [
-            pytest.param("auto", [1, 10], id="auto_max_seq"),
-            pytest.param(5, [1, 5], id="explicit_max_seq"),
-            pytest.param(None, [1, None], id="none_max_seq"),
+            pytest.param("auto", [1, 10], id="auto_max_seq"),  # dp_enabled=True -> 1, dp_enabled=False -> 10
+            pytest.param(5, [1, 5], id="explicit_max_seq"),  # dp_enabled=True -> 1, dp_enabled=False -> 5
+            pytest.param(None, [1, None], id="none_max_seq"),  # dp_enabled=True -> 1, dp_enabled=False -> None
         ],
     )
     def test_determine_max_sequences_per_example(self, sample_data, config, max_seq_input, expected_max_seq):
-        """max_sequences_per_example should be 1 for DP; for non-DP, auto→10, explicit→explicit, None→None."""
+        """Max sequences should be 1 for DP regardless of input; for non-DP, auto -> 10, explicit -> explicit value, None -> None."""
         config_copy = config.model_copy(deep=True)
         config_copy.data.max_sequences_per_example = max_seq_input
         resolver = AutoConfigResolver(sample_data, config_copy)
         result = resolver._determine_max_sequences_per_example()
+
         expected_index = 0 if config_copy.privacy.dp_enabled else 1
         assert result == {"max_sequences_per_example": expected_max_seq[expected_index]}
